NTBS-NONE: Move reference data entities to own folder
