+++ ntbs-integration-tests/NotificationPages/EpisodesPageTests.cs
@@ -9,6 +9,7 @@ using ntbs_service.Helpers;
 using ntbs_service.Models.Entities;
 using ntbs_service.Models.Enums;
 using ntbs_service.Models.FilteredSelectLists;
+using ntbs_service.Models.ReferenceEntities;
 using Xunit;
 
 namespace ntbs_integration_tests.NotificationPages
