+++ ntbs-integration-tests/NotificationPages/ManualTestResultEditPagesTests.cs
@@ -10,6 +10,7 @@ using ntbs_service.Helpers;
 using ntbs_service.Models.Entities;
 using ntbs_service.Models.Enums;
 using ntbs_service.Models.FilteredSelectLists;
+using ntbs_service.Models.ReferenceEntities;
 using ntbs_service.Models.Validations;
 using Xunit;
