+++ ntbs-service-unit-tests/Helpers/SeedingHelperTest.cs
@@ -1,5 +1,6 @@
 ﻿using ntbs_service.Helpers;
 using ntbs_service.Models.Entities;
+using ntbs_service.Models.ReferenceEntities;
 using Xunit;
 
 namespace ntbs_service_unit_tests.Helpers
