+++ ntbs-service-unit-tests/Services/UserServiceTest.cs
@@ -9,6 +9,7 @@ using Moq;
 using ntbs_service.DataAccess;
 using ntbs_service.Models.Entities;
 using ntbs_service.Models.Enums;
+using ntbs_service.Models.ReferenceEntities;
 using ntbs_service.Properties;
 using ntbs_service.Services;
 using Xunit;
