+++ ntbs-service/Models/Entities/CaseManagerTbService.cs
@@ -1,4 +1,6 @@
-﻿namespace ntbs_service.Models.Entities
+﻿using ntbs_service.Models.ReferenceEntities;
+
+namespace ntbs_service.Models.Entities
 {
     public class CaseManagerTbService
     {
