+++ ntbs-service/Models/Entities/ITravelOrVisitorDetails.cs
@@ -1,4 +1,6 @@
-﻿namespace ntbs_service.Models.Entities
+﻿using ntbs_service.Models.ReferenceEntities;
+
+namespace ntbs_service.Models.Entities
 {
     public interface ITravelOrVisitorDetails
     {
