+++ ntbs-service/Models/ReferenceEntities/Country.cs
@@ -1,4 +1,4 @@
-﻿namespace ntbs_service.Models.Entities
+﻿namespace ntbs_service.Models.ReferenceEntities
 {
     public class Country
     {
