+++ ntbs-service/Models/ReferenceEntities/Ethnicity.cs
@@ -1,4 +1,4 @@
-﻿namespace ntbs_service.Models.Entities
+﻿namespace ntbs_service.Models.ReferenceEntities
 {
     public class Ethnicity
     {
