+++ ntbs-service/Models/ReferenceEntities/ManualTestTypeSampleType.cs
@@ -1,4 +1,4 @@
-﻿namespace ntbs_service.Models.Entities
+﻿namespace ntbs_service.Models.ReferenceEntities
 {
     public class ManualTestTypeSampleType
     {
