+++ ntbs-service/Models/ReferenceEntities/Region.cs
@@ -1,4 +1,4 @@
-﻿namespace ntbs_service.Models.Entities
+﻿namespace ntbs_service.Models.ReferenceEntities
 {
     public partial class Region
     {
