+++ ntbs-service/Models/SeedData/Occupations.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using ntbs_service.Models.Entities;
+using ntbs_service.Models.ReferenceEntities;
 
 namespace ntbs_service.Models.SeedData
 {
