+++ ntbs-service/Pages/Notifications/Edit/Children/ManualTestResult.cshtml.cs
@@ -8,6 +8,7 @@ using ntbs_service.Models;
 using ntbs_service.Models.Entities;
 using ntbs_service.Models.Enums;
 using ntbs_service.Models.FilteredSelectLists;
+using ntbs_service.Models.ReferenceEntities;
 using ntbs_service.Services;
 
 namespace ntbs_service.Pages.Notifications.Edit.Children
