+++ ntbs-service/Pages/Notifications/Edit/ClinicalDetails.cshtml
@@ -4,6 +4,7 @@
 @using static NHSUK.FrontEndLibrary.TagHelpers.FormGroupType
 @using ntbs_service.Models.Entities
 @using ntbs_service.Models.Enums
+@using ntbs_service.Models.ReferenceEntities
 
 @{
     Layout = "../../Shared/_NotificationEditLayout.cshtml";
