+++ ntbs-service/Pages/Notifications/Edit/ClinicalDetails.cshtml.cs
@@ -8,6 +8,7 @@ using ntbs_service.Helpers;
 using ntbs_service.Models;
 using ntbs_service.Models.Entities;
 using ntbs_service.Models.Enums;
+using ntbs_service.Models.ReferenceEntities;
 using ntbs_service.Models.Validations;
 using ntbs_service.Services;
