+++ ntbs-service/Pages/Notifications/Edit/PatientDetails.cshtml.cs
@@ -8,6 +8,7 @@ using ntbs_service.DataAccess;
 using ntbs_service.Helpers;
 using ntbs_service.Models;
 using ntbs_service.Models.Entities;
+using ntbs_service.Models.ReferenceEntities;
 using ntbs_service.Services;
 
 namespace ntbs_service.Pages.Notifications.Edit
