+++ ntbs-service/Services/PostcodeService.cs
@@ -2,6 +2,7 @@
 using Microsoft.EntityFrameworkCore;
 using ntbs_service.DataAccess;
 using ntbs_service.Models.Entities;
+using ntbs_service.Models.ReferenceEntities;
 
 namespace ntbs_service.Services
 {
