+++ ntbs-service/Services/UserService.cs
@@ -8,6 +8,7 @@ using ntbs_service.DataAccess;
 using ntbs_service.Models;
 using ntbs_service.Models.Entities;
 using ntbs_service.Models.Enums;
+using ntbs_service.Models.ReferenceEntities;
 using ntbs_service.Properties;
 
 namespace ntbs_service.Services
